Setting up top artist pulldown
