+++ myapp/.env
@@ -61,4 +61,4 @@ SPOTIFY_ID=bcd5e565bc464208a1a03fde424733a4
 SPOTIFY_SECRET=1ec3def0425e4f85a2f3378b221c23bf
 
 # Localhost URL
-LOCALHOST_URL=http://localhost:4000
+URL=http://localhost:4000
