+++ myapp/go.mod
@@ -6,7 +6,6 @@ replace github.com/aethedigm/celeritas => ../celeritas
 
 require (
 	github.com/CloudyKit/jet/v6 v6.1.0
-	github.com/DATA-DOG/go-sqlmock v1.5.0
 	github.com/aethedigm/celeritas v0.0.0
 	github.com/go-chi/chi/v5 v5.0.7
 	github.com/go-chi/render v1.0.2
@@ -16,6 +15,7 @@ require (
 	github.com/upper/db/v4 v4.2.1
 	github.com/zmb3/spotify v1.3.0
 	golang.org/x/crypto v0.0.0-20210921155107-089bfa567519
+	golang.org/x/oauth2 v0.0.0-20211104180415-d3ed0bb246c8
 )
 
 require (
@@ -57,10 +57,8 @@ require (
 	github.com/xeipuuv/gojsonpointer v0.0.0-20180127040702-4e3ac2762d5f // indirect
 	github.com/xeipuuv/gojsonreference v0.0.0-20180127040603-bd5ef7bd5415 // indirect
 	github.com/xeipuuv/gojsonschema v1.2.0 // indirect
-	github.com/zmb3/spotify/v2 v2.3.0 // indirect
 	go.uber.org/atomic v1.7.0 // indirect
 	golang.org/x/net v0.0.0-20220225172249-27dd8689420f // indirect
-	golang.org/x/oauth2 v0.0.0-20211104180415-d3ed0bb246c8 // indirect
 	golang.org/x/sys v0.0.0-20220317061510-51cd9980dadf // indirect
 	golang.org/x/text v0.3.7 // indirect
 	google.golang.org/appengine v1.6.7 // indirect
