+++ myapp/handlers/spotify-handler.go
@@ -1,0 +1,190 @@
+package handlers
+
+import (
+	"encoding/base32"
+	"fmt"
+	"log"
+	"math/rand"
+	"myapp/data"
+	"net/http"
+	"os"
+
+	"github.com/zmb3/spotify"
+
+	"golang.org/x/oauth2"
+)
+
+var auth = spotify.Authenticator{}
+var state string
+
+var spotScopes = [2]string{spotify.ScopeUserTopRead, spotify.ScopeUserReadRecentlyPlayed}
+
+func (h *Handlers) SpotifyAuthorization(w http.ResponseWriter, r *http.Request) {
+	callback := os.Getenv("URL") + "/spotauth/callback"
+	auth = spotify.NewAuthenticator(callback, spotScopes[0], spotScopes[1])
+	randomBytes := make([]byte, 16)
+	_, err := rand.Read(randomBytes)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	state = base32.StdEncoding.WithPadding(base32.NoPadding).EncodeToString(randomBytes)
+
+	url := auth.AuthURL(state)
+	url = url + "&show_dialog=true"
+
+	http.Redirect(w, r, url, http.StatusSeeOther)
+}
+
+func (h *Handlers) SpotifyAuthorizationCallback(w http.ResponseWriter, r *http.Request) {
+	tok, err := auth.Token(state, r)
+	if err != nil {
+		http.Error(w, "Couldn't get token", http.StatusForbidden)
+		log.Fatal(err)
+	}
+	if st := r.FormValue("state"); st != state {
+		http.NotFound(w, r)
+		log.Fatalf("State mismatch: %s != %s\n", st, state)
+	}
+
+	spotclient := auth.NewClient(tok)
+	fmt.Fprintf(w, "Login Completed!")
+
+	_, err = spotclient.CurrentUser()
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	userID := h.App.Session.GetInt(r.Context(), "userID")
+	if userID == 0 || userID == -1 {
+		log.Fatal("The user_id of the current user could not be found in the session data.")
+	}
+
+	spottoken := data.SpotifyToken{
+		UserID:            userID,
+		AccessToken:       tok.AccessToken,
+		RefreshToken:      tok.RefreshToken,
+		AccessTokenExpiry: tok.Expiry,
+	}
+
+	spottoken.Upsert(spottoken)
+
+	http.Redirect(w, r, "/matches", http.StatusSeeOther)
+}
+
+func (h *Handlers) NewAccessTokenRequest(w http.ResponseWriter, r *http.Request) {
+	callback := os.Getenv("URL") + "/newspotaccesstoken/callback"
+	auth = spotify.NewAuthenticator(callback, spotScopes[0], spotScopes[1])
+	randomBytes := make([]byte, 16)
+	_, err := rand.Read(randomBytes)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	userID := h.App.Session.GetInt(r.Context(), "userID")
+	userSpotTokens, _ := h.Models.SpotifyTokens.GetSpotifyTokenForUser(userID)
+	oauth2SpotToken := oauth2.Token{
+		AccessToken:  userSpotTokens.AccessToken,
+		TokenType:    "bearer",
+		RefreshToken: userSpotTokens.RefreshToken,
+		Expiry:       userSpotTokens.AccessTokenExpiry,
+	}
+
+	client := auth.NewClient(&oauth2SpotToken)
+
+	newtoken, _ := client.Token()
+	userSpotTokens.AccessToken = newtoken.AccessToken
+	err = userSpotTokens.UpdateAccessToken(newtoken.AccessToken, newtoken.Expiry)
+	if err != nil {
+		log.Fatal(err)
+	} else {
+		fmt.Print("Spotify access token updated successfully!")
+	}
+
+	http.Redirect(w, r, "/matches", http.StatusSeeOther)
+}
+
+func (h *Handlers) NewAccessTokenAssign(w http.ResponseWriter, r *http.Request) {
+	tok, err := auth.Token(state, r)
+	if err != nil {
+		http.Error(w, "Couldn't get token", http.StatusForbidden)
+		log.Fatal(err)
+	}
+	if st := r.FormValue("state"); st != state {
+		http.NotFound(w, r)
+		log.Fatalf("State mismatch: %s != %s\n", st, state)
+	}
+
+	spotclient := auth.NewClient(tok)
+	fmt.Fprintf(w, "Login Completed!")
+
+	_, err = spotclient.CurrentUser()
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	userID := h.App.Session.GetInt(r.Context(), "userID")
+	if userID == 0 || userID == -1 {
+		log.Fatal("The user_id of the current user could not be found in the session data.")
+	}
+
+	spottoken := data.SpotifyToken{
+		UserID:            userID,
+		AccessToken:       tok.AccessToken,
+		RefreshToken:      tok.RefreshToken,
+		AccessTokenExpiry: tok.Expiry,
+	}
+
+	spottoken.Upsert(spottoken)
+
+	http.Redirect(w, r, "/matches", http.StatusSeeOther)
+}
+
+func (h *Handlers) SetSpotifyArtistsForUser(userID int) {
+	userSpotTokens, err := h.Models.SpotifyTokens.GetSpotifyTokenForUser(userID)
+	if err != nil {
+		fmt.Println("No Spotify token found for user")
+		return
+	}
+	oauth2SpotToken := oauth2.Token{
+		AccessToken:  userSpotTokens.AccessToken,
+		TokenType:    "bearer",
+		RefreshToken: userSpotTokens.RefreshToken,
+		Expiry:       userSpotTokens.AccessTokenExpiry,
+	}
+
+	auth = spotify.NewAuthenticator(os.Getenv("URL")+"/newspotaccesstoken/callback", spotScopes[0], spotScopes[1])
+	client := auth.NewClient(&oauth2SpotToken)
+
+	limit := 30
+	timeRange := "long"
+
+	artists, err := client.CurrentUsersTopArtistsOpt(&spotify.Options{
+		Limit:     &limit,
+		Timerange: &timeRange,
+	})
+	if err != nil {
+		fmt.Println("Error getting top artists for user", err)
+		return
+	}
+
+	fmt.Println("Artists added: ", artists.Artists)
+	fmt.Println("Limit:", limit)
+
+	for _, artist := range artists.Artists {
+		artistData := data.Artist{
+			SpotifyID: artist.ID.String(),
+			Name:      artist.Name,
+		}
+
+		artistData.Insert(artistData)
+	}
+
+}
+
+func min(a, b int) int {
+	if a < b {
+		return a
+	}
+	return b
+}
